\name{camelSampleTable}
\alias{camelSampleTable}
\docType{data}
\title{
ampliQueso sample data
}
\description{
Sample camel/coverage measures for 2 regions.
}
\usage{data(ampliQueso)}
\format{
  A table with 6 variables and 2 values.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(ampliQueso)
}
\keyword{datasets}
